[FIX] modify exception
