--- server/src/main/java/ssodamproject/server/common/api/RestApiException.java
@@ -1,14 +1,0 @@
-package ssodamproject.server.common.api;
-
-import lombok.Getter;
-
-@Getter
-// RESTful API에서 발생하는 예외를 나타낸다
-public class RestApiException extends RuntimeException{
-    private final ErrorType errorType;
-
-    public RestApiException(ErrorType errorType) {
-        this.errorType = errorType;
-    }
-
-}
