+++ server/src/main/java/ssodamproject/server/common/exception/RestApiException.java
@@ -5,10 +5,11 @@ import ssodamproject.server.common.enumSet.ErrorType;
 
 @Getter
 // RESTful API에서 발생하는 예외를 나타낸다
-public class RestApiException extends RuntimeException {
+public class RestApiException extends RuntimeException{
     private final ErrorType errorType;
 
     public RestApiException(ErrorType errorType) {
         this.errorType = errorType;
     }
+
 }
